+++ src/dto/form/form.hook.model.ts
@@ -3,6 +3,8 @@ import { FormHookEntity } from '../../entity/form.hook.entity'
 
 @ObjectType('FormHook')
 export class FormHookModel {
+  readonly _id: number
+
   @Field(() => ID)
   readonly id: string
 
@@ -15,8 +17,9 @@ export class FormHookModel {
   @Field({ nullable: true })
   readonly format?: string
 
-  constructor(hook: FormHookEntity) {
-    this.id = hook.id.toString()
+  constructor(id, hook: FormHookEntity) {
+    this._id = hook.id
+    this.id = id
     this.enabled = hook.enabled
     this.url = hook.url
     this.format = hook.format